# backend/openapi.yaml
openapi: 3.0.3
info:
  title: AutoFlow Runtime API
  description: |
    REST API for the AutoFlow workflow automation platform.

    ## Overview
    AutoFlow exposes two primary resource families:
    - **Workflows** — named, versioned workflow definitions (analogous to templates). Supports full CRUD.
    - **Runs** — individual executions of a workflow. Runs are created by triggering a workflow and can be monitored for status, logs, and step-level traces.

    ## WebSocket Events
    Live run log streaming is available via WebSocket at `ws://<host>/api/runs/{runId}/events`.
    The server emits newline-delimited JSON event objects (see the `RunEvent` schema).
    Connect after creating a run; the stream closes automatically when the run reaches a terminal state (`completed`, `failed`).

    ## Current Implementation Notes
    - Run storage is currently in-memory. All runs are lost on server restart (see planned PostgreSQL migration).
    - Workflow CRUD endpoints (`POST/PUT/DELETE /workflows`) persist in-memory only in v0.
    - The `/api/templates/*` routes are legacy aliases maintained for backward compatibility.
  version: "1.0.0"
  contact:
    name: AutoFlow Backend Team
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.autoflow.example.com
    description: Production
...
paths:
...
  /api/templates/{id}:
    get:
      operationId: getTemplate
      summary: Get template by ID (legacy alias for GET /api/workflows/{id})
      tags: [Templates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Template definition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Workflow not found: lead-enrichment"
...
  schemas:
...
    StepKind:
      type: string
      enum: [trigger, cron_trigger, interval_trigger, llm, transform, condition, action, output, agent, approval, mcp, file_trigger]
      description: Kind of step within a workflow
    FieldType:
      type: string
      enum: [string, number, boolean, object, "string[]", "object[]"]
      description: Data type of a config field
...
    WorkflowCategory:
      type: string
      enum: [support, sales, content, operations, marketing, finance, engineering, custom]
      description: Workflow category

    # -------------------------------------------------------------------------
    # Workflow definition schemas
    # -------------------------------------------------------------------------
    ConfigField:
      type: object
      required: [key, label, type, required]
      properties:
        key:
          type: string
          description: Machine-readable field key
          example: model
        label:
          type: string
          description: Human-readable label shown in the dashboard
          example: "LLM Model"
        type:
          $ref: "#/components/schemas/FieldType"
        required:
          type: boolean
        defaultValue:
          description: Default value for the field (any type)
        description:
          type: string
        options:
          type: array
          items:
            type: string
          description: "For string fields: restrict choices to this list"
    WorkflowStep:
      type: object
      required: [id, name, kind, description, inputKeys, outputKeys]
      properties:
        id:
          type: string
          example: step_classify
        name:
          type: string
          example: "Classify Intent"
        kind:
          $ref: "#/components/schemas/StepKind"
        description:
          type: string
        inputKeys:
          type: array
          items:
            type: string
          description: Keys from prior steps (or trigger) expected as input
        outputKeys:
          type: array
          items:
            type: string
          description: Keys produced by this step
        promptTemplate:
          type: string
          description: "LLM step: Handlebars-style prompt ({{key}} interpolation)"
        condition:
          type: string
          description: "Condition step: expression evaluated by the runtime"
          example: "intent === 'refund'"
        action:
          type: string
          description: "Action step: integration target identifier"
          example: "crm.upsertLead"
        config:
          type: object
          additionalProperties: true
          description: Step-level configuration overrides
        llmConfigId:
          type: string
          description: LLM config identifier for llm steps
        llmTier:
          type: string
          enum: [lite, standard, power]
          description: Explicit model-tier override for llm or agent steps
        mcpServerUrl:
          type: string
          description: MCP server URL for mcp steps
        mcpTool:
          type: string
          description: MCP tool name for mcp steps
        agentModel:
          type: string
          description: Model identifier for agent steps
        agentInstructions:
          type: string
          description: System instructions passed to spawned agents
        subAgentSlots:
          type: integer
          minimum: 1
          description: Parallel worker count for agent steps
        approvalAssignee:
          type: string
          description: Approval assignee for approval steps
        approvalMessage:
          type: string
          description: Approval request message shown to the approver
        approvalTimeoutMinutes:
          type: integer
          minimum: 1
          description: Timeout window for approval steps
        acceptedFileTypes:
          type: array
          items:
            type: string
          description: Allowed MIME types or extensions for file triggers
        cronExpression:
          type: string
          description: Cron expression for cron triggers
        timezone:
          type: string
          description: Timezone for cron triggers
        intervalMinutes:
          type: integer
          minimum: 1
          description: Interval frequency for interval triggers
        retry:
          $ref: "#/components/schemas/WorkflowRetryPolicy"
...
    Workflow:
      type: object
      required: [id, name, description, category, version, configFields, steps, sampleInput, expectedOutput]
      properties:
        id:
          type: string
          example: lead-enrichment
        name:
          type: string
          example: "Lead Enrichment"
        description:
          type: string
        category:
          $ref: "#/components/schemas/WorkflowCategory"
        version:
          type: string
          example: "1.0.0"
        configFields:
          type: array
          items:
            $ref: "#/components/schemas/ConfigField"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
        retry:
          $ref: "#/components/schemas/WorkflowRetryPolicy"
        errors:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
        _finally:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowStep"
        sampleInput:
          type: object
          additionalProperties: true
        expectedOutput:
          type: object
          additionalProperties: true
...
    WorkflowRetryPolicy:
      type: object
      required: [type, maxAttempts]
      properties:
        type:
          type: string
          enum: [constant, exponential, random]
        maxAttempts:
          type: integer
          minimum: 1
        maxDuration:
          type: integer
          minimum: 1
        intervalMs:
          type: integer
          minimum: 0
        delayFactor:
          type: number
          minimum: 0
          exclusiveMinimum: true
        maxInterval:
          type: integer
          minimum: 0
        warningOnRetry:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Workflow not found: lead-enrichment"